refactor: post와 관련없는 도메인 분리
