+++ src/main/java/apptive/team1/friendly/domain/comment/postcomment/entity/Comment.java
@@ -1,5 +1,6 @@
-package apptive.team1.friendly.domain.post.entity;
+package apptive.team1.friendly.domain.comment.postcomment.entity;
 
+import apptive.team1.friendly.domain.post.entity.Post;
 import apptive.team1.friendly.domain.user.data.entity.Account;
 import apptive.team1.friendly.global.baseEntity.BaseEntity;
 import lombok.AccessLevel;
