+++ src/main/java/apptive/team1/friendly/domain/comment/postcomment/service/CommentService.java
@@ -1,9 +1,9 @@
-package apptive.team1.friendly.domain.post.service;
+package apptive.team1.friendly.domain.comment.postcomment.service;
 
-import apptive.team1.friendly.domain.post.dto.CommentFormDto;
-import apptive.team1.friendly.domain.post.entity.Comment;
+import apptive.team1.friendly.domain.comment.postcomment.dto.CommentFormDto;
+import apptive.team1.friendly.domain.comment.postcomment.entity.Comment;
 import apptive.team1.friendly.domain.post.entity.Post;
-import apptive.team1.friendly.domain.post.repository.CommentRepository;
+import apptive.team1.friendly.domain.comment.postcomment.repository.CommentRepository;
 import apptive.team1.friendly.domain.post.repository.PostRepository;
 import apptive.team1.friendly.domain.user.data.entity.Account;
 import apptive.team1.friendly.domain.user.data.repository.AccountRepository;
